# use faster escapeHtml