--- a/package.json
+++ b/package.json
@@ -22,6 +22,8 @@
     "lib"
   ],
   "devDependencies": {
+    "beautify-benchmark": "~0.2.4",
+    "benchmark": "~1.0.0",
     "expect.js": "^0.3.1",
     "gulp": "^3.8.7",
     "gulp-xtemplate": "^1.2.2",