实现来自于： https://github.com/component/escape-html ，增加了 xtemplate 里面会转义的字符。

before:

```
no special characters    x 22,928,859 ops/sec ±0.98% (191 runs sampled)
single special character x  1,926,023 ops/sec ±0.83% (187 runs sampled)
many special characters  x    687,743 ops/sec ±0.98% (187 runs sampled)
```

after:

```
no special characters    x 22,687,378 ops/sec ±0.79% (192 runs sampled)
single special character x  6,571,191 ops/sec ±1.33% (190 runs sampled)
many special characters  x  3,465,183 ops/sec ±1.01% (191 runs sampled)
```