@yiminghe 2015-12-02
我感觉可以直接 require('escape-html')， 
```
`\
```

不用 escape 了，没关系的