@dead-horse 2015-12-02
rc-tools build 的时候会处理 require ?