@yiminghe 2015-12-02
build 只是为了 example , 会的，package.json 写下