@dead-horse 2015-12-02
现在只能通过 npm run pub 来生成 lib, 有没有其他的办法只生成 lib，不 publish ?