@yiminghe 2015-12-02
测试可以用 
```
./node_modules/.bin/rc-tools run compile
./node_modules/.bin/rc-tools run watch
```